+++ src/test/kotlin/com/pestphp/pest/tests/types/ThisTypeTest.kt
@@ -1,6 +1,6 @@
 package com.pestphp.pest.tests.types
 
-class ThisTypeTest: BaseTypeTest() {
+class ThisTypeTest : BaseTypeTest() {
     override fun setUp() {
         super.setUp()
 
@@ -28,6 +28,6 @@ class ThisTypeTest: BaseTypeTest() {
     }
 
     private fun assertThisCompletion() {
-        assertCompletion("expectException", "expectExceptionCode")
+        assertCompletion("expectException", "expectExceptionCode", "someProtectedMethod", "someStaticMethod", "protectedField")
     }
 }