+++ src/test/resources/com/pestphp/pest/PestUtil/PestTestFunctionCallWithNamesapce.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Unit;
+
+it('basic', function () {
+    $this->assertTrue(true);
+});