+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Added support for navigation between tests and test subject ([#53](https://github.com/pestphp/pest-intellij/pull/53))
 - Added error reporting to GitHub issues ([#55](https://github.com/pestphp/pest-intellij/pull/55))
 - Added inspection for duplicate test names in same file. ([#56](https://github.com/pestphp/pest-intellij/pull/56))
+- Added completions for static and protected $this methods. ([#66](https://github.com/pestphp/pest-intellij/pull/66))
+- Added completions $this fields declared in beforeEach functions. ([#66](https://github.com/pestphp/pest-intellij/pull/66))
 - Added pcov coverage engine support
 
 ### Changed
@@ -18,6 +20,8 @@
 ### Removed
 
 ### Fixed
+- Fixed duplicate test name error when no test name is given yet. ([#61](https://github.com/pestphp/pest-intellij/pull/61))
+- Fixed finding tests with namespace at the top. ([#65](https://github.com/pestphp/pest-intellij/pull/65))
 
 ### Security
 ## [0.3.3]