+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ## [Unreleased]
 ### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+## [0.4.0]
+### Added
 - Added support for showing pest version ([#52](https://github.com/pestphp/pest-intellij/pull/52))
 - Type provider for Pest test functions ([#48](https://github.com/pestphp/pest-intellij/pull/48))
 - Added support for navigation between tests and test subject ([#53](https://github.com/pestphp/pest-intellij/pull/53))